requirements.txt
Jinja2==3.1.6
PyYAML==6.0.2
boto3==1.36.26
botocore==1.36.26
cryptography==44.0.2
fastapi==0.115.12
google-api-python-client==2.172.0
google-auth==2.40.1
google-auth-httplib2==0.2.0
httplib2==0.22.0
protobuf==5.29.4
pydantic==2.11.3
pydantic-settings==2.8.1
python-dotenv==1.1.0
python-multipart==0.0.20
requests==2.32.3
rsa==4.8
s3transfer==0.11.0
starlette==0.46.2
typing-inspection==0.4.0
typing_extensions==4.13.2
uvicorn==0.34.2
